requirements.txt
streamlit
streamlit-lottie
Pillow
st-pages
streamlit_authenticator
deta